+++ wallet_app/packages/wallet_core/lib/src/models/version_state.dart
@@ -18,7 +18,7 @@ sealed class FlutterVersionState with _$FlutterVersionState {
   const factory FlutterVersionState.notify() = FlutterVersionState_Notify;
   const factory FlutterVersionState.recommend() = FlutterVersionState_Recommend;
   const factory FlutterVersionState.warn({
-    required BigInt expiresInSeconds,
+    required int expiresInSeconds,
   }) = FlutterVersionState_Warn;
   const factory FlutterVersionState.block() = FlutterVersionState_Block;
 }
